Updated navbar tests
