+++ testcafe/pages/navBar.ts
@@ -1,6 +1,6 @@
 import { ReactSelector } from "testcafe-react-selectors";
 
-export default class NavBar {
+export default class Navbar {
   navBar = ReactSelector("NavBar");
   home = this.navBar.find("#homeLink");
   createPoll = this.navBar.find("#createPollLink");
